requirements.txt
# Core dependencies
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# Database
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# HTTP requests for task handlers
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
responses==0.24.1
fakeredis==2.20.1

# Development
python-dotenv==1.0.0

# Type hints
typing-extensions==4.9.0
